+++ src/collection.rs
@@ -3,7 +3,6 @@
 
 mod cereal;
 mod insomnia;
-mod jetbrains_env;
 mod models;
 mod recipe_tree;
 mod rest_format;